+++ frontend/src/components/Section/Section.test.js
@@ -1,7 +1,7 @@
 import { describe, expect, it, test } from "vitest";
 import { render } from "@testing-library/svelte";
 
-import SectionTest from "./Section.test.svelte";
+import SectionTest from "./SectionTest.svelte";
 
 
 describe("Section", () => {
