+++ frontend/src/layouts/Story.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from './Layout.astro';
-import Section from '../components/Section.svelte';
+import Section from '../components/Section/Section.svelte';
 import Title from '../components/Title.svelte';
 import Panel from '../components/dashboard/Panel.svelte';
 import Link from "../components/Link.svelte";
