+++ frontend/src/pages/consultations/[consultationId]/index.astro
@@ -2,7 +2,7 @@
 import Layout from "../../../layouts/Layout.astro";
 
 import ConsultationDetail from "../../../components/screens/ConsultationDetail.svelte";
-import Section from "../../../components/Section.svelte";
+import Section from "../../../components/Section/Section.svelte";
 
 const { consultationId } = Astro.params;
 ---
