+++ frontend/src/pages/consultations/[consultationId]/responses/[responseId].astro
@@ -1,6 +1,6 @@
 ---
 import Layout from "../../../../layouts/Layout.astro";
-import Section from "../../../../components/Section.svelte";
+import Section from "../../../../components/Section/Section.svelte";
 
 import QuestionDetail from "../../../../components/screens/QuestionDetail.svelte";
