+++ frontend/src/pages/consultations/index.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../../layouts/Layout.astro';
-import Section from '../../components/Section.svelte';
+import Section from '../../components/Section/Section.svelte';
 import Title from '../../components/Title.svelte';
 
 import ConsultationList from '../../components/screens/ConsultationList.svelte';
